+++ backend/libraries/group_chat_core/src/lib.rs
@@ -214,7 +214,8 @@ impl GroupChatCore {
                         updates.role_changed = true;
                     }
                 }
-                ChatEventInternal::ParticipantsAdded(_)
+                ChatEventInternal::MembersAddedToDefaultChannel(_)
+                | ChatEventInternal::ParticipantsAdded(_)
                 | ChatEventInternal::ParticipantsRemoved(_)
                 | ChatEventInternal::ParticipantJoined(_)
                 | ChatEventInternal::ParticipantLeft(_)
@@ -324,6 +325,11 @@ impl GroupChatCore {
                 }
                 ChatEventInternal::ParticipantLeft(p) => {
                     user_updates_handler.mark_member_updated(&mut result, p.user_id, true);
+                }
+                ChatEventInternal::MembersAddedToDefaultChannel(m) => {
+                    for user_id in m.user_ids.iter() {
+                        user_updates_handler.mark_member_updated(&mut result, *user_id, false);
+                    }
                 }
                 ChatEventInternal::RoleChanged(rc) => {
                     for user_id in rc.user_ids.iter() {